[FARM-432]Added general button for present modals in playground
